+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.45.0
 ### Features
 * add `id` option that allow to create same terminal using hot reload [#978](https://github.com/jcubic/jquery.terminal/issues/978)
+* allow using Object URLs in links [#982](https://github.com/jcubic/jquery.terminal/issues/982)
 ### Bugfix
 * fix `terminal::login()` when user already authenticated [#980](https://github.com/jcubic/jquery.terminal/issues/980)